[stable] [analyzer] Clear watcher subscriptions to avoid leaking package:watcher data
Analyzer adds watchers to all root folders it has opened.
When editing, say, a pubspec.yaml file it ~starts over, cancels the
old subscriptions and creates new ones. It keeps references to the old
subscriptions too though.

This in turn leaks data.

On Linux internally in package:watcher it clears some stuff as it's
cancelled, but it doesn't on Windows, nor on Mac.

The means it can leak an infinite amount of `_Entry` classes,
and for instance in https://github.com/dart-lang/sdk/issues/52447 a
user has 40+ million of those classes.

Reproducing this I got to 40+ million too, at which point the
"Retained size" of all the `_Entry` objects was 7.39GB.

This fixes the leak by clearing the list.

Cherry-pick: https://dart-review.googlesource.com/c/sdk/+/308940
Change-Id: I8a4bdbc97259bd8e13209909769330ca6d81b2a5
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/311466
Commit-Queue: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>